requirements.txt
Django>=3.2,<5.0
django-ipware
ipinfo
celery
redis
django-celery-beat
django-ratelimit
djangorestframework==3.15.1
